Move extension related test helpers
We use the naming "extension" rather than "bookingExtension" for consistency with arrivals, departures, etc.
