+++ cypress.config.ts
@@ -4,7 +4,7 @@ import auth from './integration_tests/mockApis/auth'
 import tokenVerification from './integration_tests/mockApis/tokenVerification'
 import premises from './integration_tests/mockApis/premises'
 import booking from './integration_tests/mockApis/booking'
-import bookingExtension from './integration_tests/mockApis/bookingExtension'
+import extension from './integration_tests/mockApis/extension'
 import arrival from './integration_tests/mockApis/arrival'
 import nonArrival from './integration_tests/mockApis/nonArrival'
 import departure from './integration_tests/mockApis/departure'
@@ -36,7 +36,7 @@ export default defineConfig({
         ...tokenVerification,
         ...premises,
         ...booking,
-        ...bookingExtension,
+        ...extension,
         ...departure,
         ...cancellation,
         ...confirmation,
